V0.25

+++ routes/shepherd/kickstart.js
@@ -7,37 +7,46 @@ module.exports = (shepherd) => {
    *  params: coin, type
    */
   shepherd.get('/kick', (req, res, next) => {
-    const _coin = req.query.coin;
-    const _keepWallet = req.query.keepwallet;
+    if (shepherd.checkToken(req.query.token)) {
+      const _coin = req.query.coin;
+      const _keepWallet = req.query.keepwallet;
 
-    if (!_coin) {
+      if (!_coin) {
+        const errorObj = {
+          msg: 'error',
+          result: 'no coin name provided',
+        };
+
+        res.end(JSON.stringify(errorObj));
+      } else {
+        const _location = path.join(_coin === 'KMD' ? shepherd.komodoDir : `${shepherd.komodoDir}/${_coin}`);
+
+        if (fs.existsSync(_location)) {
+          const items = fs.readdirSync(_location);
+
+          for (let i = 0; i < items.length; i++) {
+            if (items[i].indexOf('wallet.dat') === -1) {
+              fs.removeSync(`${_location}/${items[i]}`);
+            } else if (!_keepWallet) {
+              fs.removeSync(`${_location}/${items[i]}`);
+            }
+          }
+        }
+
+        const successObj = {
+          msg: 'success',
+          result: `${_coin} native is kicked`,
+        };
+
+        res.end(JSON.stringify(successObj));
+      }
+    } else {
       const errorObj = {
         msg: 'error',
-        result: 'no coin name provided',
+        result: 'unauthorized access',
       };
 
       res.end(JSON.stringify(errorObj));
-    } else {
-      const _location = path.join(_coin === 'KMD' ? shepherd.komodoDir : `${shepherd.komodoDir}/${_coin}`);
-
-      if (fs.existsSync(_location)) {
-        const items = fs.readdirSync(_location);
-
-        for (let i = 0; i < items.length; i++) {
-          if (items[i].indexOf('wallet.dat') === -1) {
-            fs.removeSync(`${_location}/${items[i]}`);
-          } else if (!_keepWallet) {
-            fs.removeSync(`${_location}/${items[i]}`);
-          }
-        }
-      }
-
-      const successObj = {
-        msg: 'success',
-        result: `${_coin} native is kicked`,
-      };
-
-      res.end(JSON.stringify(successObj));
     }
   });
